refactor(backend): move client schema in client app
Author: Rudy Savini - @SavGRY
