+++ backend/core/schema.py
@@ -4,26 +4,6 @@ from typing import Literal
 from pydantic import BaseModel
 
 
-class Client(BaseModel):
-    id: int
-    name: str
-    code: int
-    address: str
-    city: str
-    province: str
-
-    class Config:
-        orm_mode = True
-
-
-class ClientFromForm(BaseModel):
-    code: int
-    name: str
-    address: str
-    city: str
-    province: str
-
-
 class Document(BaseModel):
     id: int
     creation_date: datetime
